requirements.txt
tkinter
pillow
pyinstaller